Added BDT workflow files
BDT training and evaluation is outsourced to the ROOT TMVA library and
is intended to be used on a per-PFP evaluation basis.

`bdt_prep` handles training a TMVA method based on the output TTrees
produced by the `examples/create_bdt_trees.py` script.

`bdt_apply` takes training weights and an analysis file. It then
evaluates a BDT score for each PFP and writes that additional branch to
the analysis file.